+++ packages/bridge-controller/src/bridge-controller.ts
@@ -997,6 +997,7 @@ export class BridgeController extends StaticIntervalPollingController<BridgePoll
         };
       }
       case UnifiedSwapBridgeEventName.AssetDetailTooltipClicked:
+      case UnifiedSwapBridgeEventName.AssetPickerOpened:
         return baseProperties;
       // These events may be published after the bridge-controller state is reset
       // So the BridgeStatusController populates all the properties