+++ packages/bridge-controller/src/utils/metrics/types.ts
@@ -218,6 +218,9 @@ type RequiredEventContextFromClientBase = {
     failures: string[];
     refresh_count: number;
   };
+  [UnifiedSwapBridgeEventName.AssetPickerOpened]: {
+    asset_location: 'source' | 'destination';
+  };
   [UnifiedSwapBridgeEventName.PollingStatusUpdated]: TradeData &
     Pick<QuoteFetchData, 'price_impact'> &
     Omit<RequestMetadata, 'security_warnings'> &
@@ -298,6 +301,7 @@ export type EventPropertiesFromControllerState = {
     refresh_count: number;
   };
   [UnifiedSwapBridgeEventName.StatusValidationFailed]: RequestParams;
+  [UnifiedSwapBridgeEventName.AssetPickerOpened]: null;
   [UnifiedSwapBridgeEventName.PollingStatusUpdated]: null;
 };